created jwa.ts and moved some functions
